Im updating my branch

+++ menu2.php
@@ -1,3 +1,6 @@
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.0/jquery.min.js"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+
 <div class="menu" id="menu">
 	<ul>
 		<li><a href="dashboard.php">Home</a></li>
@@ -5,13 +8,16 @@
         <li><a href="#">Manage Shifts</a><b aria-haspopup="true" aria-controls="p1"></b>
         	<ul id="p1">
 				<li><a href="#"> Shift Types/Descriptions</a></li>
-				<li><a href="#"> Add Shifts</a></li>
-				<li><a href="#"> Manage Existing Shifts</a></li>
+				<li><a href="createShifts.php"> Add Shifts</a></li>
+				<li><a href="shifts.php"> Manage Existing Shifts</a></li>
 				</ul>
         </li>
 		<li><a href="#">Manage Volunteers</a></li>
-		<li><a href="/scripts/edit/deleteEvent.php">Delete Event</a></li>
+		<li><a href="/scripts/edit/deleteEvent.php" onclick="return confirm('Are you sure you want to delete this event? ');">Delete Event</a></li>
 		<li><a href="./scripts/config/logout.php">Logout</a></li>
 
 	</ul>
 </div>
+
+
+href="eventOverview.php?id=<?php echo $row['idEvent'];?>
